Cleanup addExtension
