+++ platform/cli/src/commands/utils/writeExtensionToConfig.js
@@ -1,0 +1,21 @@
+import readPluginConfigFile from './readPluginConfigFile.js';
+import { addExtensionToConfig } from './manipulatePluginConfigFile.js';
+import writePluginConfigFile from './writePluginConfigFile.js';
+
+export default function writeExtensionToConfig(packageName, yarnInfo) {
+  const installedVersion = yarnInfo.version;
+  const pluginConfig = readPluginConfigFile();
+
+  if (!pluginConfig) {
+    pluginConfig = {
+      extensions: [],
+      modes: [],
+    };
+  }
+
+  addExtensionToConfig(pluginConfig, {
+    packageName,
+    version: installedVersion,
+  });
+  writePluginConfigFile(pluginConfig);
+}
