+++ platform/viewer/package.json
@@ -49,7 +49,7 @@
   },
   "dependencies": {
     "@babel/runtime": "7.7.6",
-    "@ohif-test/extension-clock": "1.0.2",
+    "@ohif-test/extension-clock": "^1.0.2",
     "@ohif-test/mode-clock": "^1.0.5",
     "@ohif/core": "^2.5.1",
     "@ohif/extension-cornerstone": "^2.4.0",
