+++ platform/viewer/pluginConfig.json
@@ -17,6 +17,9 @@
             "version": "1.0.1"
         },
         {
+            "version": "4.12.11"
+        },
+        {
             "packageName": "@ohif-test/extension-clock",
             "version": "1.0.2"
         }
